update input interface for users

+++ xe-scrapper.py
@@ -24,31 +24,38 @@ def getExchangeRate(br,date,fromCurrency,toCurrency):
             return float(rate)
 
 if __name__== '__main__': # this need to be later converted i a function
-    starting_year = "2000"
-    ending_year = "2003"
+    starting_year = raw_input("Input the starting year ")
+    ending_year = raw_input("Input the ending year ")
+    sys.stdout.write("\nThe starting year {}\n".format(starting_year))
+    sys.stdout.write("\n The ending year {}\n".format(ending_year))
 
     outDir = os.path.join( os.getcwd(),'xeDatabase') # dir for all inputs and outputs
     if not os.path.exists(outDir):
         os.makedirs(outDir)
-
+    print("\n")
     br = initMechanize()
     start = datetime.datetime.strptime("01-01-"+starting_year,"%d-%m-%Y")
     end = datetime.datetime.strptime("01-01-"+ending_year,"%d-%m-%Y")
     date_generated = [start + datetime.timedelta(days=x) for x in range(0, (end-start).days)]
+    checkFileExist = os.path.join(outDir,starting_year+"-"+ending_year+".p")
 
-    exchange_dict = dict()
-    index = []
-    for idx,date in enumerate(date_generated):
-        index.append(idx)
-        rate = getExchangeRate(br,date.strftime("%Y-%m-%d"),"myr","twd")
-        sys.stdout.write( 'Scraping: {}... '.format(date.strftime("%Y-%m-%d")) )
-        if "date" in exchange_dict:
-            exchange_dict["date"].append(date.strftime("%Y-%m-%d"))
-            exchange_dict["rates"].append(rate)
-        else:
-            exchange_dict["date"] = [date.strftime("%Y-%m-%d")]
-            exchange_dict["rates"] = [rate]
-        time.sleep(5)
-        sys.stdout.write( '\n' )
-    df = pd.DataFrame(data=exchange_dict,index = index)
-    df.to_pickle(outDir+"/"+starting_year+"-"+ending_year+".p")
+    if not os.path.exists(checkFileExist):
+        exchange_dict = dict()
+        index = []
+        for idx,date in enumerate(date_generated):
+            index.append(idx)
+            rate = getExchangeRate(br,date.strftime("%Y-%m-%d"),"myr","twd")
+            print 'Scraping: {}... \r'.format(date.strftime("%Y-%m-%d")),
+            sys.stdout.write("\033[F")
+            if "date" in exchange_dict:
+                exchange_dict["date"].append(date.strftime("%Y-%m-%d"))
+                exchange_dict["rates"].append(rate)
+            else:
+                exchange_dict["date"] = [date.strftime("%Y-%m-%d")]
+                exchange_dict["rates"] = [rate]
+            time.sleep(5)
+            sys.stdout.write( '\n' )
+        df = pd.DataFrame(data=exchange_dict,index = index)
+        df.to_pickle(outDir+"/"+starting_year+"-"+ending_year+".p")
+    else:
+        sys.stdout.write('File already exist')
